test: extract chips tests suite into separate file (#6692)
